+++ ChangeLog.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Python class generation.
  - Perl class generation.
 
-## [0.3.0] - 2019-11-16
+## [0.3.0] - 2019-11-18
 ### Added
  - Examples of basic usage and features.
  - Option to give a class-level include list.
@@ -25,6 +25,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Struct members may have default values specified to allow for the generation
    of a default constructor.
  - Support for virtual functions.
+ - Pointer wrapper classes are by default given a constructor that takes a
+   pointer to the equivalent struct and wraps it in the class.
 
 ### Fixed
  - `bin/wrapture` is now executable.